Fri, Jul 7, 2023, 10:23 AM +06:00

+++ bluesheet/18.c
@@ -2,7 +2,7 @@
 
 int main(void)
 {
-    unsigned long arr[];
+    int arr[];
 
     return 0;
 }
